python_testing/requirements.txt
asciitable==0.8.0
mysql-connector-python==8.0.27
protobuf==3.19.1
